requirements.txt
aiohttp
discord.py
requests
BismuthClient
